refactor Misc.kt
