+++ app/src/main/java/com/example/tonetuner_v2/audio/audioProcessing/AudioProc.kt
@@ -11,7 +11,7 @@ import com.example.tonetuner_v2.app.AppModel.PROC_BUFFER_SIZE
 import com.example.tonetuner_v2.app.AppModel.QUALITY_QUEUE_SIZE
 import com.example.tonetuner_v2.audio.audioSources.AudioSource
 import com.example.tonetuner_v2.ui.navigation.MainLayout
-import com.example.tonetuner_v2.util.forcedOffer
+import com.example.tonetuner_v2.extensions.forcedOffer
 import java.util.concurrent.ArrayBlockingQueue
 import java.util.concurrent.BlockingQueue
 /**
