+++ app/src/main/java/com/example/tonetuner_v2/ui/composables/old/GaugeTuner.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.drawscope.rotate
 import com.example.signallib.enums.Note
-import com.example.tonetuner_v2.util.toString
+import com.example.tonetuner_v2.extensions.toString
 import kotlin.math.ln
 
 @Composable
